# root/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/subscription-plans": {
      "post": {
        "operationId": "SubscriptionPlansController_create",
        "summary": "Create Subscription Plan",
        "description": "Adds a price tier to EnerTrack's catalogue. Adding a tier is a data change, not a code change: the billing engine reads every pricing knob off this row.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for any write.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created successfully."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "409": {
            "description": "A plan with that name already exists."
          }
        },
        "tags": [
          "subscription-plans"
        ]
      },
      "get": {
        "operationId": "SubscriptionPlansController_findAll",
        "summary": "List All Subscription Plans",
        "description": "The full catalogue including inactive tiers and every pricing knob. Not tenant-scoped: the catalogue is global, so a client sees the same list EnerTrack staff do.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for any write.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of all plan records."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          }
        },
        "tags": [
          "subscription-plans"
        ]
      }
    },
...
    "/api/subscription-plans/{id}": {
      "get": {
        "operationId": "SubscriptionPlansController_findOne",
        "summary": "Get Subscription Plan by ID",
        "description": "Retrieves a single price tier by its plan_id.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for any write.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan record returned."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Plan with the given ID not found."
          }
        },
        "tags": [
          "subscription-plans"
        ]
      },
      "put": {
        "operationId": "SubscriptionPlansController_put",
        "summary": "Replace Subscription Plan",
        "description": "Completely replaces a price tier. Send a full body.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for any write.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutSubscriptionPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan replaced successfully."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Plan with the given ID not found."
          }
        },
        "tags": [
          "subscription-plans"
        ]
      },
      "patch": {
        "operationId": "SubscriptionPlansController_update",
        "summary": "Update Subscription Plan",
        "description": "Partially updates a tier — a price change, or retiring it with is_active: false. Takes effect on the next invoice generated, with no redeploy.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for any write.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan updated successfully."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Plan with the given ID not found."
          }
        },
        "tags": [
          "subscription-plans"
        ]
      },
      "delete": {
        "operationId": "SubscriptionPlansController_remove",
        "summary": "Delete Subscription Plan",
        "description": "Permanently removes a tier. Refused while any subscription still references it, since the billing engine would then fail to resolve a plan_id.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for any write.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan deleted successfully."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Plan with the given ID not found."
          },
          "409": {
            "description": "Plan is still in use by a subscription."
          }
        },
        "tags": [
          "subscription-plans"
        ]
      }
    },
    "/api/subscription-plans/{id}/subscribers": {
      "get": {
        "operationId": "SubscriptionPlansController_getSubscribers",
        "summary": "Get Plan Subscribers",
        "description": "Organisations currently contracted on this tier. Used by the Super Admin revenue console to show plan distribution.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for any write.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of subscriber summaries."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          }
        },
        "tags": [
          "subscription-plans"
        ]
      }
    },
...
    "/api/subscriptions/{id}": {
      "get": {
        "operationId": "SubscriptionsController_findOne",
        "summary": "Get Subscription by ID",
        "description": "Retrieves a single contract by its subscription_id.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-org-id",
            "in": "header",
            "description": "Organisation the caller is acting within. Omitted by EnerTrack staff, who get the cross-tenant view.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for writes; a client's own roles may read.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription record returned."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Subscription with the given ID not found."
          }
        },
        "tags": [
          "subscriptions"
        ]
      },
      "put": {
        "operationId": "SubscriptionsController_put",
        "summary": "Replace Subscription",
        "description": "Completely replaces a contract. Send a full body.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for writes; a client's own roles may read.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription replaced successfully."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Subscription with the given ID not found."
          }
        },
        "tags": [
          "subscriptions"
        ]
      },
      "patch": {
        "operationId": "SubscriptionsController_update",
        "summary": "Update Subscription",
        "description": "Partially updates a contract, for example moving it from trial to active. organization_id cannot be changed.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for writes; a client's own roles may read.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription updated successfully."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Subscription with the given ID not found."
          }
        },
        "tags": [
          "subscriptions"
        ]
      },
      "delete": {
        "operationId": "SubscriptionsController_remove",
        "summary": "Delete Subscription",
        "description": "Permanently removes a contract. Refused once any platform invoice references it — cancel it instead.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC. Super Admin for writes; a client's own roles may read.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription deleted successfully."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Subscription with the given ID not found."
          },
          "409": {
            "description": "Invoices still reference this subscription."
          }
        },
        "tags": [
          "subscriptions"
        ]
      }
    },
...
  },
  "info": {
    "title": "EnerTrack API",
    "description": "Backend API for EnerTrack — Facility Management, Sustainability & Finance Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      },
      "x-org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-org-id"
      }
    },
    "schemas": {
...
      "CreateSubscriptionPlanDto": {
        "type": "object",
        "properties": {}
      },
      "PutSubscriptionPlanDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSubscriptionPlanDto": {
        "type": "object",
        "properties": {}
      },
...
      "PutSubscriptionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSubscriptionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}